Fixed drag functionality and lines
Added so that each link has a line representing the link. Also fixed the
drag functionality so that it actually works. Still doesn't update the
children tho, but that's in the pipe.
Most importantly, i fixed self.scale which was set to
dictWargs['height'] for some reason causing the width of the objects to
be scaled up totally wrong.

In this particular case the scale was width*width == 10*10 == 100. So
the click-check was off by 90 pixels!